+++ tests/render/renderer/test_suite_scaffold_notebook_renderer.py
@@ -16,9 +16,7 @@ def test_render_snapshot_test(titanic_data_context):
     csv_path = batch_kwargs["path"]
     suite_name = "my_suite"
     suite = titanic_data_context.create_expectation_suite(suite_name)
-    renderer = SuiteScaffoldNotebookRenderer(
-        titanic_data_context, suite, batch_kwargs
-    )
+    renderer = SuiteScaffoldNotebookRenderer(titanic_data_context, suite, batch_kwargs)
     obs = renderer.render(None)
     assert isinstance(obs, nbformat.NotebookNode)
     ## NOTE!!! - When updating this snapshot be sure to include the dynamic
@@ -30,7 +28,12 @@ def test_render_snapshot_test(titanic_data_context):
         "cells": [
             {
                 "cell_type": "markdown",
-                "source": "# Scaffold a new Expectation Suite (BETA)\nUse this notebook to scaffold a new expectations suite. This process helps you\navoid writing lots of boilerplate when authoring suites.\n\n**Expectation Suite Name**: `my_suite`\n\nWe'd love it if you **reach out to us on** the [**Great Expectations Slack Channel**](https://greatexpectations.io/slack)",
+                "source": """# Scaffold a new Expectation Suite (BETA)
+This process helps you avoid writing lots of boilerplate when authoring suites by allowing you to select columns you care about and letting a profiler write some candidate expectations for you to adjust.
+
+**Expectation Suite Name**: `bar`
+
+We'd love it if you **reach out to us on** the [**Great Expectations Slack Channel**](https://greatexpectations.io/slack)""",
                 "metadata": {},
             },
             {
@@ -44,7 +47,13 @@ def test_render_snapshot_test(titanic_data_context):
             },
             {
                 "cell_type": "markdown",
-                "source": "## Select the columns you want to scaffold expectations on\n\nSimply uncomment columns that are important. You can select multiple lines and\nuse a jupyter keyboard shortcut to toggle each line: **Linux/Windows**:\n`Ctrl-/`, **macOS**: `Cmd-/`",
+                "source": """## Select the columns you want to scaffold expectations on
+
+Great Expectations will choose which expectations might make sense for a column based on the **data type** and **cardinality** of the data in each selected column.
+
+Simply uncomment columns that are important. You can select multiple lines and
+use a jupyter keyboard shortcut to toggle each line: **Linux/Windows**:
+`Ctrl-/`, **macOS**: `Cmd-/`""",
                 "metadata": {},
             },
             {
@@ -56,7 +65,18 @@ def test_render_snapshot_test(titanic_data_context):
             },
             {
                 "cell_type": "markdown",
-                "source": "## Run the scaffolder\n\nThis is highly configurable depending on your goals. You can include or exclude\ncolumns, and include or exclude expectation types (when applicable). [The \nExpectation Glossary](http://docs.greatexpectations.io/en/latest/expectation_glossary.html) \ncontains a list of possible expectations.\n\nNote that the profiler is not very smart, so it does it's best to decide on\napplicability.\n\n**To get to a production grade suite, you should [edit this \nsuite](http://docs.greatexpectations.io/en/latest/command_line.html#great-expectations-suite-edit) \nafter this scaffold gets you close to what you want.**",
+                "source": """## Run the scaffolder
+
+The suites generated here are **not meant to be production suites** - they are **scaffolds to build upon**.
+
+**To get to a production grade suite, will definitely want to [edit this 
+suite](http://docs.greatexpectations.io/en/latest/command_line.html#great-expectations-suite-edit) 
+after scaffolding gets you close to what you want.**
+
+This is highly configurable depending on your goals. You can include or exclude
+columns, and include or exclude expectation types (when applicable). [The 
+Expectation Glossary](http://docs.greatexpectations.io/en/latest/expectation_glossary.html) 
+contains a list of possible expectations.""",
                 "metadata": {},
             },
             {
@@ -140,9 +160,7 @@ def test_notebook_execution_with_pandas_backend(titanic_data_context):
     assert not os.path.isfile(notebook_path)
 
     # Create notebook
-    renderer = SuiteScaffoldNotebookRenderer(
-        titanic_data_context, suite, batch_kwargs
-    )
+    renderer = SuiteScaffoldNotebookRenderer(titanic_data_context, suite, batch_kwargs)
     renderer.render_to_disk(notebook_path)
     assert os.path.isfile(notebook_path)