+++ tests/test_utils.py
@@ -587,7 +587,7 @@ def is_library_installed(library_name):
     try:
         importlib.import_module(library_name)
         return True
-    except ModuleNotFoundError as e:
+    except ModuleNotFoundError:
         return False
 
 
@@ -597,4 +597,3 @@ def safe_remove(path):
             os.remove(path)
         except OSError as e:
             print(e)
-            pass